chore(db): restructure providers crate
