+++ crates/net/rpc/src/eth/api/mod.rs
@@ -1,8 +1,8 @@
 //! Provides everything related to `eth_` namespace
 
 use reth_interfaces::Result;
-use reth_primitives::U64;
-use reth_provider::{BlockProvider, ChainInfo, StateProviderFactory};
+use reth_primitives::{ChainInfo, U64};
+use reth_provider::{BlockProvider, StateProviderFactory};
 use reth_rpc_types::Transaction;
 use reth_transaction_pool::TransactionPool;
 use std::sync::Arc;
