+++ crates/primitives/src/chain/spec.rs
@@ -11,7 +11,7 @@ use std::collections::{BTreeMap, HashMap};
 /// The Etereum mainnet spec
 pub static MAINNET: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
     chain: Chain::mainnet(),
-    genesis: serde_json::from_str(include_str!("../res/genesis/mainnet.json"))
+    genesis: serde_json::from_str(include_str!("../../res/genesis/mainnet.json"))
         .expect("Can't deserialize Mainnet genesis json"),
     genesis_hash: H256(hex!("d4e56740f876aef8c010b86a40d5f56745a118d0906a34e69aec8c0db1cb8fa3")),
     hardforks: BTreeMap::from([
@@ -39,7 +39,7 @@ pub static MAINNET: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
 /// The Goerli spec
 pub static GOERLI: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
     chain: Chain::goerli(),
-    genesis: serde_json::from_str(include_str!("../res/genesis/goerli.json"))
+    genesis: serde_json::from_str(include_str!("../../res/genesis/goerli.json"))
         .expect("Can't deserialize Goerli genesis json"),
     genesis_hash: H256(hex!("bf7e331f7f7c1dd2e05159666b3bf8bc7a8a3a9eb1d518969eab529dd9b88c1a")),
     hardforks: BTreeMap::from([
@@ -56,7 +56,7 @@ pub static GOERLI: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
 /// The Sepolia spec
 pub static SEPOLIA: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
     chain: Chain::sepolia(),
-    genesis: serde_json::from_str(include_str!("../res/genesis/sepolia.json"))
+    genesis: serde_json::from_str(include_str!("../../res/genesis/sepolia.json"))
         .expect("Can't deserialize Sepolia genesis json"),
     genesis_hash: H256(hex!("25a5cc106eea7138acab33231d7160d69cb777ee0c2c553fcddf5138993e6dd9")),
     hardforks: BTreeMap::from([(Hardfork::MergeNetsplit, 1735371)]),
