+++ crates/storage/provider/Cargo.toml
@@ -8,28 +8,33 @@ readme = "README.md"
 description = "Reth storage provider."
 
 [dependencies]
+#reth
 reth-codecs = { path = "../codecs" }
 reth-primitives = { path = "../../primitives" }
 reth-interfaces = { path = "../../interfaces" }
 reth-rpc-types = { path = "../../net/rpc-types" }
 reth-db = { path = "../db" }
+
+# codecs
+serde = { version = "1.0.*", default-features = false }
+postcard = { version = "1.0.2", features = ["alloc"] }
+parity-scale-codec = { version = "3.2.1", features = ["bytes"] }
+
+# misc
 async-trait = "0.1.57"
 thiserror = "1.0.37"
 auto_impl = "1.0"
 tokio = { version = "1.21.2", features = ["sync"] }
 bytes = "1.2"
-
-# codecs
-serde = { version = "1.0.*", default-features = false }
-postcard = { version = "1.0.2", features = ["alloc"] }
-heapless = "0.7.16"
-parity-scale-codec = { version = "3.2.1", features = ["bytes"] }
 futures = "0.3.25"
 tokio-stream = "0.1.11"
 rand = "0.8.5"
+modular-bitfield = "0.11.2"
+heapless = "0.7.16"
+
+# feature test-utils
 arbitrary = { version = "1.1.7", features = ["derive"], optional = true }
 secp256k1 = { version = "0.24.2", default-features = false, features = ["alloc", "recovery", "rand"], optional = true }
-modular-bitfield = "0.11.2"
 parking_lot = { version = "0.12", optional = true }
 
 [dev-dependencies]
