+++ crates/storage/provider/src/traits/state.rs
@@ -1,17 +1,8 @@
+use super::AccountProvider;
+use crate::BlockHashProvider;
 use auto_impl::auto_impl;
 use reth_interfaces::Result;
-use reth_primitives::{
-    Account, Address, BlockHash, BlockNumber, Bytes, StorageKey, StorageValue, H256,
-};
-
-use crate::block::BlockHashProvider;
-
-/// Account provider
-#[auto_impl(&)]
-pub trait AccountProvider: Send + Sync {
-    /// Get basic account information.
-    fn basic_account(&self, address: Address) -> Result<Option<Account>>;
-}
+use reth_primitives::{Address, BlockHash, BlockNumber, Bytes, StorageKey, StorageValue, H256};
 
 /// Function needed for executor.
 #[auto_impl(&)]
